+++ app/components/Notification/Notification.tsx
@@ -1,7 +1,9 @@
 import clsx from 'clsx';
 import React, { PropsWithChildren } from 'react';
-import CloseIcon from '@resources/svg/cross.svg';
 import { motion, Transition, Variants } from 'framer-motion';
+
+import CloseIcon from '@resources/svg/cross.svg';
+import { ClassName } from '~/utils/types';
 import Button from '~/cross-app-components/Button';
 import { NotificationLevel } from '~/features/NotificationSlice';
 
@@ -22,10 +24,12 @@ const transitionVariants: Variants = {
     },
 };
 
-type NotificationProps = PropsWithChildren<{
-    level: NotificationLevel;
-    onCloseClick(): void;
-}>;
+type NotificationProps = ClassName &
+    PropsWithChildren<{
+        level: NotificationLevel;
+        disableClose?: boolean;
+        onCloseClick(): void;
+    }>;
 
 const modifier: Partial<Record<NotificationLevel, string>> = {
     [NotificationLevel.Error]: styles.error,
@@ -34,20 +38,24 @@ const modifier: Partial<Record<NotificationLevel, string>> = {
 export default function Notification({
     children,
     level,
+    disableClose = false,
     onCloseClick,
+    className,
 }: NotificationProps) {
     return (
         <motion.div
-            className={clsx(styles.root, modifier[level])}
+            className={clsx(styles.root, modifier[level], className)}
             transition={transition}
             variants={transitionVariants}
             initial="initial"
             animate="enter"
             exit="exit"
         >
-            <Button clear className={styles.close} onClick={onCloseClick}>
-                <CloseIcon width="11" />
-            </Button>
+            {disableClose || (
+                <Button clear className={styles.close} onClick={onCloseClick}>
+                    <CloseIcon width="11" />
+                </Button>
+            )}
             {children}
         </motion.div>
     );